+++ packages/patrol/CHANGELOG.md
@@ -21,6 +21,7 @@
   - accept downloads
 - Add support for Patrol extensions. (#3160)
 - Bump ktor packages. (#3187)
+- Fix incorrect test description in the failure log entry when a test fails in `patrol develop` (hot restart) mode. (#3167)
 
 ## 4.7.1